requirements.txt
dnspython
prettytable
python-ldap
PyYAML